requirements.txt
langchain-community 
langchain 
streamlit 
sentence-transformers 
ctransformers 
ipykernel 
python-box
mysql-connector-python